requirements.txt
gensim==4.2.0
treamlit==1.11.0
